% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello}
\usage{
hello(x = 2, y = 3)
}
\arguments{
\item{x}{An integer}

\item{y}{An integer}
}
\value{
A chracter vector
}
\description{
Hello world
}
\details{
Hello world details
}
\examples{
hello()
}
