% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello2}
\alias{hello2}
\title{Title}
\usage{
hello2(...)
}
\arguments{
\item{...}{Arguments passed on to \code{hello}
\describe{
  \item{x}{An integer}
  \item{y}{An integer}
}}
}
\value{
A chracter vector
}
\description{
Title
}
